Combined the tilesets in stage_road1.tscn
The three tilesets in stage_road1.tscn only contained
one tile each. I've taken these three tilesets and
replaced them with a single tileset that contains
all the tiles in one.

This results in a prettier node hierarchy in the
scene as well as making editing easier if the
layout of tiles needs to be changed.